# swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "Free Mentors is a social initiative where accomplished professionals become role models to young people to provide free mentorship sessions",
		"version": "1.0.0",
		"title": "FREE MENTORS",
		"termsOfService": "http://swagger.io/terms/",
		"contact": {
			"email": "mujohn26@gmail.com"
		},
		"license": {
			"name": "MIT",
			"url": "http://www.apache.org/licenses/LICENSE-2.0.html"
		}
	},
	"host": "freementor26.herokuapp.com",
	"basePath": "/api/v1",
...
	"schemes": [

		"http"
	],
	"paths": {
...
		"/mentors/{id}": {
			"get": {
				"tags": [
					"Mentor"
				],
				"summary": "return a specific mentor",
				"description": "at this endpoint should return one mentor of the given id",
				"operationId": "specificmentor",
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "id of the mentor",
						"required": true,
						"type": "integer"

					},
					{
						"name": "x-auth-token",
						"in": "header",
						"description": "the token of the user to view that mentor",
						"required": true,
						"type": "string"

					}
				],
				"responses": {
					"200": {
						"description": "a mentor",
						"schema":{
							"type":"object"
						}

					},
					"401": {
						"description": "you are not an admin"
					},
					"404": {
						"description": "not have access to server"
					},
					"500": {
						"description": "server not found"
					}
				}
				
			}
		},
...
	},
...
}